refactor: rename files and update dependencies
