+++ src/resources/tasks/task.service.js
@@ -1,4 +1,4 @@
-const tasksRepo = require('./task.memory.repository');
+const tasksRepo = require('./task.db.repository');
 
 const getAll = boardId => tasksRepo.getAll(boardId);
