+++ src/resources/users/user.service.js
@@ -1,4 +1,4 @@
-const usersRepo = require('./user.memory.repository');
+const usersRepo = require('./user.db.repository');
 
 const getAll = () => usersRepo.getAll();
